Move IP ranges data source into domain package
