+++ internal/provider/iprange/data_source_test.go
@@ -1,4 +1,4 @@
-package provider
+package iprange
 
 import (
 	"context"
